requirements.txt
appdirs==1.4.4
attrs==24.2.0
beautifulsoup4==4.12.3
behave==1.2.6
blinker==1.8.2
Brotli==1.1.0
bs4==0.0.2
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
cssselect==1.2.0
DataRecorder==3.5.3
docx2txt==0.8
DownloadKit==2.0.2
et-xmlfile==1.1.0
exceptiongroup==1.2.2
execnet==2.1.1
fake-useragent==1.5.1
fasteners==0.19
filelock==3.15.4
Flask==3.0.3
FlowViewer==0.3.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.7
importlib_metadata==8.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kaitaistruct==0.10
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.1.0
opencv-python==4.10.0.84
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
parameterized==0.9.0
parse==1.20.2
parse-type==0.6.2
pdbp==1.5.4
pip-review==1.3.0
platformdirs==4.2.2
pluggy==1.5.0
pminit==0.8.0
py==1.11.0
pyasn1==0.6.0
pychrome==0.2.4
pycparser==2.22
pydivert==2.1.0
pydub==0.25.1
pyee==11.1.0
Pygments==2.18.0
pyjsparser==2.7.1
pynose==1.5.2
pyOpenSSL==24.2.1
pyotp==2.9.0
pyparsing==3.1.2
PyPasser==0.0.5
pyppeteer==2.0.0
pyquery==2.0.0
pyreadline3==3.4.1
PySocks==1.7.1
pytest==8.3.2
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==14.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.7.1
sbvirtualdisplay==1.3.0
selenium==4.23.1
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.10.4
tabcompleter==1.3.3
tldextract==5.1.2
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.19
w3lib==2.2.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.4
wsproto==1.2.0
zipp==3.20.0
zstandard==0.23.0
